feat: serve frontend build from backend
